# apps/backend-api/src/modules/resource/resource.controller.ts
import {
  Body,
  Controller,
  Get,
  HttpCode,
  Param,
  ParseIntPipe,
  Post,
  Query,
  StreamableFile,
  UploadedFile,
  UseInterceptors,
} from '@nestjs/common'
import { FileInterceptor } from '@nestjs/platform-express'
import { ResourceService } from '@/modules/resource/resource.service'
import { ResourcePageRespDto } from '@/modules/resource/dto/resource.page.resp.dto'
import { ResourcePageDto } from '@/modules/resource/dto/resource.page.dto'
import { ResourcePageOptionDto } from '@/modules/resource/dto/resource.page.option.dto'
import { ResourceRespDto } from '@/modules/resource/dto/resource.resp.dto'
import { ResourceListRespDto } from '@/modules/resource/dto/resource.list.resp.dto'
import { ResourceCreateDto } from '@/modules/resource/dto/resource.create.dto'
import {
  ResourceBatchUpdateDto,
  ResourceUpdateDto,
} from '@/modules/resource/dto/resource.update.dto'
import { PaginatedResult } from '@/dto/pagination-response.dto'
import { ApiConsumes, ApiOkResponse, ApiOperation, ApiProduces } from '@nestjs/swagger'
import {
  ApiSuccessListResponse,
  ApiSuccessPageResponse,
  ApiSuccessResponse,
} from '@/decorators/api-response.decorator'
import { BatchDto, BatchRespDto, BatchUpdateStatusDto } from '@/dto/batch.dto'
import { UpdateStatusDto } from '@/dto/update-status.dto'
import { CurrentUser } from '@/decorators/current-user.decorator'
import { CurrentLoginResponseDto } from '../auth/auth.dto'
import { ListResp } from '@/dto/base.dto'
import { RequirePermissions } from '@/decorators/require-permissions.decorator'
import { PERMISSIONS } from '@repo/shared'
import { EXCEL_CONTENT_TYPE, ExcelUploadFile } from '@/modules/excel/excel.types'
import { ImportResultDto } from '@/modules/excel/dto/import-result.dto'

@Controller('resource')
export class ResourceController {
  constructor(private readonly resourceService: ResourceService) {}

  @Get('page')
  @RequirePermissions(PERMISSIONS.SYS_RESOURCE_LIST_PAGE)
  @ApiOperation({ summary: '分页查询资源列表' })
  @ApiSuccessPageResponse(ResourcePageRespDto)
  async pageResource(
    @Query() query: ResourcePageDto,
  ): Promise<PaginatedResult<ResourcePageRespDto>> {
    return await this.resourceService.pageResource(query)
  }

  @Get('list')
  @RequirePermissions(PERMISSIONS.SYS_RESOURCE_LIST_PAGE)
  @ApiOperation({ summary: '查询全量资源列表' })
  @ApiSuccessPageResponse(ResourcePageRespDto)
  async listAllResource(@Query() query: ResourcePageDto): Promise<ListResp<ResourcePageRespDto>> {
    return await this.resourceService.listAllResource(query)
  }

  @Get('listByUser')
  @RequirePermissions(PERMISSIONS.SYS_RESOURCE_LIST_PAGE)
  @ApiOperation({ summary: '查询用户的菜单资源' })
  @ApiSuccessListResponse(ResourcePageRespDto)
  async listResourceByUser(
    @Query('id', new ParseIntPipe({ optional: true })) id: number | undefined,
    @Query('types') types: string | undefined,
    @Query('notInMenu', new ParseIntPipe({ optional: true })) notInMenu: number | undefined,
    @CurrentUser() user: CurrentLoginResponseDto,
  ): Promise<ListResp<ResourcePageRespDto>> {
    return await this.resourceService.listByUser(id || user.id, user.roleCodes, types, notInMenu)
  }

  @Post('schema-generator-holder-page-resource')
  @HttpCode(200)
  @ApiOperation({ summary: '请勿调用，用于生成前端DTO', deprecated: true })
  async _pageResource(@Body() _body: ResourcePageDto) {
    return null
  }

  @Get('option')
  @RequirePermissions(PERMISSIONS.SYS_RESOURCE_LIST_PAGE)
  @ApiOperation({ summary: '分页查询资源列表(可选字段)' })
  @ApiSuccessPageResponse(ResourcePageRespDto)
  async pageOptionResource(
    @Query() query: ResourcePageOptionDto,
  ): Promise<PaginatedResult<Partial<ResourcePageRespDto>>> {
    return await this.resourceService.pageOptionResource(query)
  }

  @Post('schema-generator-holder-page-option-resource')
  @HttpCode(200)
  @ApiOperation({ summary: '请勿调用，用于生成前端DTO', deprecated: true })
  async _pageOptionResource(@Body() _body: ResourcePageOptionDto) {
    return null
  }

  @Get('find/:id')
  @RequirePermissions(PERMISSIONS.SYS_RESOURCE_LIST_VIEW)
  @ApiOperation({ summary: '按id查询资源' })
  @ApiSuccessResponse(ResourceRespDto)
  async findResourceById(@Param('id', ParseIntPipe) id: number): Promise<ResourceRespDto | null> {
    const resource = await this.resourceService.findResourceById(id)
    return resource
  }

  @Get('batch/query')
  @RequirePermissions(PERMISSIONS.SYS_RESOURCE_LIST_VIEW)
  @ApiOperation({ summary: '按ids查询资源' })
  @ApiSuccessResponse(ResourceListRespDto)
  async findResourceListByIds(@Query() query: BatchDto): Promise<ResourceListRespDto | null> {
    return await this.resourceService.findResourceListByIds(query.ids)
  }

  @Post('create')
  @RequirePermissions(PERMISSIONS.SYS_RESOURCE_LIST_CREATE)
  @ApiOperation({ summary: '创建资源' })
  @ApiSuccessResponse(ResourceRespDto)
  async createResource(@Body() body: ResourceCreateDto): Promise<ResourceRespDto | null> {
    return await this.resourceService.createResource(body)
  }

  @Post('update/:id')
  @RequirePermissions(PERMISSIONS.SYS_RESOURCE_LIST_EDIT)
  @ApiOperation({ summary: '修改资源' })
  @ApiSuccessResponse(ResourceRespDto)
  async updateResource(
    @Param('id', ParseIntPipe) id: number,
    @Body() body: ResourceUpdateDto,
  ): Promise<ResourceRespDto | null> {
    return await this.resourceService.updateResource(id, body)
  }

  @Post('batchUpdate')
  @RequirePermissions(PERMISSIONS.SYS_RESOURCE_LIST_EDIT)
  @ApiOperation({ summary: '批量修改资源' })
  @ApiSuccessResponse()
  async batchUpdateResource(
    @Body() body: ResourceBatchUpdateDto,
  ): Promise<{ info?: string | null }> {
    return await this.resourceService.batchUpdateResource(body.list)
  }

  @Post('delete/:id')
  @RequirePermissions(PERMISSIONS.SYS_RESOURCE_LIST_DELETE)
  @ApiOperation({ summary: '删除资源' })
  @ApiSuccessResponse()
  async removeResource(@Param('id') id: number) {
    return await this.resourceService.removeResource(id)
  }

  @Post('batch/delete')
  @RequirePermissions(PERMISSIONS.SYS_RESOURCE_LIST_DELETE)
  @ApiOperation({ summary: '批量删除资源' })
  @ApiSuccessResponse(BatchRespDto)
  async batchRemoveResource(@Body() body: BatchDto): Promise<BatchRespDto | null> {
    return await this.resourceService.batchRemoveResource(body.ids)
  }

  @Post('updateStatus/:id')
  @RequirePermissions(PERMISSIONS.SYS_RESOURCE_LIST_EDIT)
  @ApiOperation({ summary: '修改资源状态' })
  @ApiSuccessResponse()
  async updateResourceStatus(@Param('id') id: number, @Body() body: UpdateStatusDto) {
    return await this.resourceService.updateResourceStatus(id, body)
  }

  @Post('batch/status')
  @RequirePermissions(PERMISSIONS.SYS_RESOURCE_LIST_EDIT)
  @ApiOperation({ summary: '批量修改资源状态' })
  @ApiSuccessResponse(BatchRespDto)
  async batchUpdateResourceStatus(
    @Body() body: BatchUpdateStatusDto,
  ): Promise<BatchRespDto | null> {
    return await this.resourceService.batchUpdateResourceStatus(body)
  }
  @Post('resetSort')
  @RequirePermissions(PERMISSIONS.SYS_RESOURCE_LIST_EDIT)
  @ApiOperation({ summary: '重置列表顺序' })
  @ApiSuccessResponse()
  async resetResourceListSort(): Promise<null> {
    return await this.resourceService.resetResourceListSort()
  }

  @Post('template')
  @RequirePermissions(PERMISSIONS.SYS_RESOURCE_LIST_IMPORT)
  @ApiOperation({ summary: '下载导入资源模板' })
  @ApiProduces(EXCEL_CONTENT_TYPE)
  async downloadResourceTemplate(): Promise<StreamableFile> {
    return await this.resourceService.downloadTemplate()
  }

  @Post('import')
  @RequirePermissions(PERMISSIONS.SYS_RESOURCE_LIST_IMPORT)
  @ApiConsumes('multipart/form-data')
  @UseInterceptors(FileInterceptor('file'))
  @ApiOperation({ summary: '导入资源数据' })
  @ApiOkResponse({ type: ImportResultDto })
  async importResource(
    @UploadedFile() file: ExcelUploadFile | undefined,
  ): Promise<ImportResultDto> {
    return await this.resourceService.importResource(file)
  }

  @Post('export')
  @RequirePermissions(PERMISSIONS.SYS_RESOURCE_LIST_EXPORT)
  @ApiOperation({ summary: '导出资源数据' })
  @ApiProduces(EXCEL_CONTENT_TYPE)
  async exportResource(@Query() query: ResourcePageDto): Promise<StreamableFile> {
    return await this.resourceService.exportResource(query)
  }
}