# packages/types/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/resource/page": {
      "get": {
        "operationId": "ResourceController_pageResource",
        "parameters": [
          {
            "name": "label",
            "required": false,
            "in": "query",
            "description": "资源名称",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "uniqueProp",
            "required": false,
            "in": "query",
            "description": "唯一编码",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "parentUniqueProp",
            "required": false,
            "in": "query",
            "description": "唯一父编码",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "状态：0-禁用，1-启用",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "资源类型：0-目录 1-页面 2-按钮 3-列",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "notInMenu",
            "required": false,
            "in": "query",
            "description": "是否在菜单中显示：0-显示，1-不显示",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sortNumber",
            "required": false,
            "in": "query",
            "description": "排序号",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "menuPath",
            "required": false,
            "in": "query",
            "description": "菜单路径",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "createdAt",
            "required": false,
            "in": "query",
            "description": "创建时间",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "updatedAt",
            "required": false,
            "in": "query",
            "description": "修改时间",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deletedAt",
            "required": false,
            "in": "query",
            "description": "删除时间",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "createdAtStart",
            "required": false,
            "in": "query",
            "description": "创建时间开始",
            "schema": {
              "example": "2000-10-10 00:00:00",
              "type": "string"
            }
          },
          {
            "name": "createdAtEnd",
            "required": false,
            "in": "query",
            "description": "创建时间结束",
            "schema": {
              "example": "2000-10-10 23:59:59",
              "type": "string"
            }
          },
          {
            "name": "updatedAtStart",
            "required": false,
            "in": "query",
            "description": "修改时间开始",
            "schema": {
              "example": "2000-10-10 00:00:00",
              "type": "string"
            }
          },
          {
            "name": "updatedAtEnd",
            "required": false,
            "in": "query",
            "description": "修改时间结束",
            "schema": {
              "example": "2000-10-10 23:59:59",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "分页成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "allOf": [
                            {
                              "$ref": "#/components/schemas/PaginatedResult"
                            },
                            {
                              "required": ["total", "page", "pageSize", "list"],
                              "properties": {
                                "total": {
                                  "type": "number",
                                  "description": "总条数",
                                  "example": 100
                                },
                                "page": {
                                  "type": "number",
                                  "description": "当前页码",
                                  "example": 1
                                },
                                "pageSize": {
                                  "type": "number",
                                  "description": "每页条数",
                                  "example": 10
                                },
                                "list": {
                                  "type": "array",
                                  "items": {
                                    "$ref": "#/components/schemas/ResourcePageRespDto"
                                  }
                                }
                              }
                            }
                          ]
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "分页查询资源列表",
        "tags": ["Resource"]
      }
    },
    "/resource/list": {
      "get": {
        "operationId": "ResourceController_listAllResource",
        "parameters": [
          {
            "name": "label",
            "required": false,
            "in": "query",
            "description": "资源名称",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "uniqueProp",
            "required": false,
            "in": "query",
            "description": "唯一编码",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "parentUniqueProp",
            "required": false,
            "in": "query",
            "description": "唯一父编码",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "状态：0-禁用，1-启用",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "资源类型：0-目录 1-页面 2-按钮 3-列",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "notInMenu",
            "required": false,
            "in": "query",
            "description": "是否在菜单中显示：0-显示，1-不显示",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sortNumber",
            "required": false,
            "in": "query",
            "description": "排序号",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "menuPath",
            "required": false,
            "in": "query",
            "description": "菜单路径",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "createdAt",
            "required": false,
            "in": "query",
            "description": "创建时间",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "updatedAt",
            "required": false,
            "in": "query",
            "description": "修改时间",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deletedAt",
            "required": false,
            "in": "query",
            "description": "删除时间",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "createdAtStart",
            "required": false,
            "in": "query",
            "description": "创建时间开始",
            "schema": {
              "example": "2000-10-10 00:00:00",
              "type": "string"
            }
          },
          {
            "name": "createdAtEnd",
            "required": false,
            "in": "query",
            "description": "创建时间结束",
            "schema": {
              "example": "2000-10-10 23:59:59",
              "type": "string"
            }
          },
          {
            "name": "updatedAtStart",
            "required": false,
            "in": "query",
            "description": "修改时间开始",
            "schema": {
              "example": "2000-10-10 00:00:00",
              "type": "string"
            }
          },
          {
            "name": "updatedAtEnd",
            "required": false,
            "in": "query",
            "description": "修改时间结束",
            "schema": {
              "example": "2000-10-10 23:59:59",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "分页成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "allOf": [
                            {
                              "$ref": "#/components/schemas/PaginatedResult"
                            },
                            {
                              "required": ["total", "page", "pageSize", "list"],
                              "properties": {
                                "total": {
                                  "type": "number",
                                  "description": "总条数",
                                  "example": 100
                                },
                                "page": {
                                  "type": "number",
                                  "description": "当前页码",
                                  "example": 1
                                },
                                "pageSize": {
                                  "type": "number",
                                  "description": "每页条数",
                                  "example": 10
                                },
                                "list": {
                                  "type": "array",
                                  "items": {
                                    "$ref": "#/components/schemas/ResourcePageRespDto"
                                  }
                                }
                              }
                            }
                          ]
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "查询全量资源列表",
        "tags": ["Resource"]
      }
    },
...
    "/resource/schema-generator-holder-page-resource": {
      "post": {
        "deprecated": true,
        "operationId": "ResourceController__pageResource",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourcePageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "请勿调用，用于生成前端DTO",
        "tags": ["Resource"]
      }
    },
    "/resource/option": {
      "get": {
        "operationId": "ResourceController_pageOptionResource",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "keyword",
            "required": false,
            "in": "query",
            "description": "关键字",
            "schema": {
              "example": "",
              "type": "string"
            }
          },
          {
            "name": "fields",
            "required": false,
            "in": "query",
            "description": "返回的列表字段",
            "schema": {
              "example": "",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "分页成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "allOf": [
                            {
                              "$ref": "#/components/schemas/PaginatedResult"
                            },
                            {
                              "required": ["total", "page", "pageSize", "list"],
                              "properties": {
                                "total": {
                                  "type": "number",
                                  "description": "总条数",
                                  "example": 100
                                },
                                "page": {
                                  "type": "number",
                                  "description": "当前页码",
                                  "example": 1
                                },
                                "pageSize": {
                                  "type": "number",
                                  "description": "每页条数",
                                  "example": 10
                                },
                                "list": {
                                  "type": "array",
                                  "items": {
                                    "$ref": "#/components/schemas/ResourcePageRespDto"
                                  }
                                }
                              }
                            }
                          ]
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "分页查询资源列表(可选字段)",
        "tags": ["Resource"]
      }
    },
    "/resource/schema-generator-holder-page-option-resource": {
      "post": {
        "deprecated": true,
        "operationId": "ResourceController__pageOptionResource",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourcePageOptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "请勿调用，用于生成前端DTO",
        "tags": ["Resource"]
      }
    },
    "/resource/find/{id}": {
      "get": {
        "operationId": "ResourceController_findResourceById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ResourceRespDto"
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "按id查询资源",
        "tags": ["Resource"]
      }
    },
    "/resource/batch/query": {
      "get": {
        "operationId": "ResourceController_findResourceListByIds",
        "parameters": [
          {
            "name": "ids",
            "required": true,
            "in": "query",
            "description": "用“,”拼接的id",
            "schema": {
              "example": "",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ResourceListRespDto"
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "按ids查询资源",
        "tags": ["Resource"]
      }
    },
    "/resource/create": {
      "post": {
        "operationId": "ResourceController_createResource",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ResourceRespDto"
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "创建资源",
        "tags": ["Resource"]
      }
    },
    "/resource/update/{id}": {
      "post": {
        "operationId": "ResourceController_updateResource",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ResourceRespDto"
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "修改资源",
        "tags": ["Resource"]
      }
    },
    "/resource/batchUpdate": {
      "post": {
        "operationId": "ResourceController_batchUpdateResource",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceBatchUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "default": null,
                          "nullable": true
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "批量修改资源",
        "tags": ["Resource"]
      }
    },
    "/resource/delete/{id}": {
      "post": {
        "operationId": "ResourceController_removeResource",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "default": null,
                          "nullable": true
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "删除资源",
        "tags": ["Resource"]
      }
    },
    "/resource/batch/delete": {
      "post": {
        "operationId": "ResourceController_batchRemoveResource",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/BatchRespDto"
                        }
                      },
                      "required": ["data"]
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "批量删除资源",
        "tags": ["Resource"]
      }
    },
...
  },
  "info": {
    "title": "BACKED API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "head": {
            "$ref": "#/components/schemas/ResponseHeadDto"
          }
        },
        "required": ["head"]
      },
...
      "PaginatedResult": {
        "type": "object",
        "properties": {}
      },
...
      "ResourcePageRespDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "资源id"
          },
          "label": {
            "type": "string",
            "description": "资源名称"
          },
          "uniqueProp": {
            "type": "string",
            "description": "唯一编码"
          },
          "parentUniqueProp": {
            "type": "string",
            "description": "唯一父编码"
          },
          "status": {
            "type": "number",
            "description": "状态：0-禁用，1-启用"
          },
          "type": {
            "type": "number",
            "description": "资源类型：0-目录 1-页面 2-按钮 3-列"
          },
          "menuPath": {
            "type": "string",
            "description": "菜单路径"
          },
          "notInMenu": {
            "type": "number",
            "description": "是否在菜单中显示：0-显示，1-不显示"
          },
          "sortNumber": {
            "type": "number",
            "description": "排序号"
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间"
          },
          "updatedAt": {
            "type": "string",
            "description": "修改时间"
          },
          "deletedAt": {
            "type": "string",
            "description": "删除时间"
          }
        },
        "required": ["id"]
      },
...
      "ResourcePageOptionDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "description": "页码"
          },
          "pageSize": {
            "type": "number",
            "description": "每页条数"
          },
          "keyword": {
            "type": "string",
            "description": "关键字",
            "example": ""
          },
          "fields": {
            "type": "string",
            "description": "返回的列表字段",
            "example": ""
          }
        }
      },
...
      "ResourceBatchUpdateDto": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResourcePartialDto"
            }
          }
        },
        "required": ["list"]
      },
...
    }
  }
}